+++ src/_modules/root/root.controller.ts
@@ -5,13 +5,10 @@ class RootController {
   constructor(){}
 
   async helloWorld({set, user, session, log}: any) {
-    log.error(user, "User");
-    log.info(session, "Session");
-
     try {
-      console.log(user);
-      console.log(session);
-      console.log(user?.profileId ?? null);
+      console.log("User: ",user);
+      console.log("Session: ",session);
+      console.log("ProfileID: ",user?.profileId ?? null);
       
       return { message: `Welcome to the ${consts.server.name} Server! Version ${consts.server.version}` }
     } catch (error) {
@@ -24,8 +21,9 @@ class RootController {
 
   async helloTime({ user, session, params, store:{timezone} }: any) {
 
-    console.warn(user);
-    console.warn(session);
+    console.log("User: ",user);
+    console.log("Session: ",session);
+    console.log("ProfileID: ",user?.profileId ?? null);
 
     const currentHour = new Date().getHours();